Update URLs in Webmap feature collections (#5)
Added code in PortalContentPost.py update_webmaps function to replace
the host names in the webmap json. This is basically a "catch-all" to
replace any other references to the host name that were not replaced in
the other code blocks in this function which specifically extract the
"url" JSON key values outside of feature collections. This code was
specifically added to address the cases where publishers had include csv
content into the webmap that had URL references.