Fix to #1428
PortalContentPost.py failing with "Invalid group capability" error.

The issue is that the createGroup endpoint in the portal sharing api
only accepts form-data and not json. So the capabilities and tags list
properties had to be converted to comma delimited strings.

Also modified the PortalContentPost.py script to print out the original
group id so that debugging issues would be easier. Added an addition
if/else block around the call to the reassign_group function to only
call if the group was created.